LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-size-nested-invalidation-expected.txt


FAIL Queries that skip a container are invalidated correctly assert_equals: expected "true" but got ""
